% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engel.R
\docType{data}
\name{Engel Data}
\alias{Engel Data}
\alias{engel}
\title{Engel Data}
\format{A data frame containing 235 observations on 2 variables
\describe{
\item{income}{annual household income in Belgian francs}
\item{foordexp}{annual household food expenditure in Belgian francs}
}}
\usage{
data(engel)
}
\description{
Engel food expenditure data used in Koenker and Bassett(1982). This is a regression data set consisting of 235 observations on income and expenditure on food for Belgian working class households.
}
\references{
Koenker, R. and Bassett, G (1982) Robust Tests of Heteroscedasticity based on Regression Quantiles; Econometrica 50, 43-61.
}
\keyword{datasets}
