% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trust.R
\docType{data}
\name{Trust Data}
\alias{Trust Data}
\alias{trust}
\title{Trust Data}
\format{A data frame containing 180 observations for 12 countries. Data taken from the Eurobarometer, The Hertie School of Governance. 
\describe{
\item{Trust_in_the_ECB}{Trust in the European Central Bank}
\item{Trust_in_the_EU}{Trust in the European Union}
\item{countryname}{countryname identifier}
\item{year}{year identifier}
}}
\usage{
data(trust)
}
\description{
Data on 12 European Countries
}
\keyword{datasets}
