% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urqb.R
\name{urqb}
\alias{urqb}
\title{Unconditional Quantile Regression}
\usage{
urqb(data,tau,formula,kernel=NULL,cluster=cluster)
}
\arguments{
\item{data}{a data.frame in which to interpret the variables named in the formula}

\item{tau}{the quantile(s) to be estimated, this must be a number (or a vector of numbers) strictly between 0 and 1.}

\item{formula}{a formula object, with the response on the left of a ~ operator, and the terms, separated by + operators, on the right.}

\item{kernel}{a character string giving the smoothing kernel to be used. This must match one of "gaussian", "rectangular", "triangular", "epanechnikov", "biweight", "cosine" or "optcosine", with default "gaussian".}

\item{cluster}{column name of variable to be used in order to obtain cluster robust standard errors.}
}
\description{
Function Not intended for user. Returns an object of class "urq" that represents an Unconditional Quantile Regression Fit.
}
\examples{
NULL
}
\seealso{
\code{\link{density},\link{urq}}
}

